fix: resolve compiler warnings in fs.c, buffer.c, and compat.h
- Add (void)mode to silence unused variable warning on Windows in mkdirSync
- Add S_IFLNK and S_ISLNK compat macros for platforms missing them
- Use uintptr_t instead of unsigned long in ANT_PTR macro to fix pointer-to-int cast warnings on 64-bit Windows

+++ include/ant.h
@@ -14,7 +14,7 @@
 #define STR_PROTO_LEN 9
 
 #define ANT_STRING(s)         js_mkstr(js, s, sizeof(s) - 1)
-#define ANT_PTR(ptr)          js_mknum((unsigned long)(ptr))
+#define ANT_PTR(ptr)          js_mknum((uintptr_t)(ptr))
 #define ANT_COPY(buf, len, s) cpy(buf, len, s, sizeof(s) - 1)
 
 struct js;
